+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 Select a Spreadsheet Light version below to view the changelog history:
 
-- [Spreadsheet Light 0.2.0](#V0.2.4) **Current**
+- [Spreadsheet Light 0.2.0](#V0.2.7) **Current**
+- [Spreadsheet Light 0.2.0](#V0.2.6)
+- [Spreadsheet Light 0.2.0](#V0.2.5)
+- [Spreadsheet Light 0.2.0](#V0.2.4)
 - [Spreadsheet Light 0.2.0](#V0.2.3)
 - [Spreadsheet Light 0.2.0](#V0.2.2)
 - [Spreadsheet Light 0.2.0](#V0.2.1)
@@ -12,6 +15,19 @@ Select a Spreadsheet Light version below to view the changelog history:
 - [Spreadsheet Light 0.1.9](#V0.1.9)
 - [Spreadsheet Light 0.1.7](#V0.1.7)
 - [Spreadsheet Light 0.1.6](#V0.1.6)
+
+## V0.2.7
+
+- Adding `reducer` function to `Spreadsheet` class so now the data can be transformed to another custom type line by line
+- Adding to readme the use of `toSafeString` more context
+
+## V0.2.6
+
+- Adding logic to `toSafeString` function to change back `empty` values to `zero string` value
+
+## V0.2.5
+
+- Adding missing `changed` flag when a `Spreadsheet` object updates are applied
 
 ## V0.2.4